requirements.txt
hdx-python-api~=5.9.9
beautifulsoup4~=4.11.2
geopandas~=0.12.2
pandas~=1.5.3
rasterstats~=0.16.0
